// src/app/models/gift.model.ts
export class GiftModel {
    id: number = null;
    name: string = null;
    dateIncluse: Date = null;
    responsible: string = null;
    quantity: number = null;
    entrance: number = null;
    exit: number = null;

    dateIncluseExport: string = null;
    totalExport: number = null;
    files: any = null;
}